+++ packages/terra-core-docs/src/terra-dev-site/doc/section-header/example/AccordionSectionHeader.jsx
@@ -1,0 +1,22 @@
+import React, { useState } from 'react';
+import SectionHeaderExampleTemplate from './SectionHeaderExampleTemplate';
+
+const AccordionSectionHeader = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleClick = () => {
+    setIsOpen(!isOpen);
+  };
+
+  const sectionHeaderProps = {
+    title: 'I can accordion, click me',
+    isOpen,
+    onClick: handleClick,
+  };
+
+  return (
+    <SectionHeaderExampleTemplate title="Accordion Section Header" exampleProps={sectionHeaderProps} />
+  );
+};
+
+export default AccordionSectionHeader;