+++ packages/terra-core-docs/src/terra-dev-site/test/status-view/StatusViewNotAuthorized.test.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import StatusView from 'terra-status-view';
+
+export default () => (
+  <StatusView id="statusView" variant="not-authorized" />
+);