+++ packages/terra-show-hide/tests/wdio/show-hide-spec.js
@@ -1,6 +1,6 @@
 Terra.describeViewports('Show Hide', ['medium'], () => {
   describe('Default', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/default-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/default-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('default collapsed');
@@ -53,7 +53,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('Custom button text', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/custom-button-text-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/custom-button-text-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('custom button collapsed');
@@ -69,7 +69,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('Initially open', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/initially-open-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/initially-open-show-hide'));
 
     it('should display expanded Show Hide', () => {
       Terra.validates.element('initially open expanded');
@@ -77,7 +77,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('No preview', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/no-preview-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/no-preview-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('no preview collapsed');
@@ -85,7 +85,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('Button align start', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/button-align-start-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/button-align-start-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('button align start collapsed');
@@ -101,7 +101,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('Button align center', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/button-align-center-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/button-align-center-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('button align center collapsed');
@@ -117,7 +117,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
   });
 
   describe('Button align end', () => {
-    before(() => browser.url('/raw/tests/terra-show-hide/show-hide/button-align-end-show-hide'));
+    before(() => browser.url('/raw/tests/cerner-terra-core-docs/show-hide/button-align-end-show-hide'));
 
     it('should display collapsed Show Hide', () => {
       Terra.validates.element('button align end collapsed');
@@ -134,7 +134,7 @@ Terra.describeViewports('Show Hide', ['medium'], () => {
 
   describe('Long button text', () => {
     before(() => {
-      browser.url('/raw/tests/terra-show-hide/show-hide/long-button-text-show-hide');
+      browser.url('/raw/tests/cerner-terra-core-docs/show-hide/long-button-text-show-hide');
       $('#root').moveTo({ xOffset: 0, yOffset: 700 }); // move cursor to remove focus styles
     });