+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased
 
 * Added
+  * Migrate docs from terra-section-header.
+  * Migrate docs from terra-spacer.
+  * Migrate docs from terra-signature.
+  * Migrate docs from terra-show-hide.
+  * Migrate docs from terra-status-view.
+  * Migrate docs from terra-search-field.
+  * Migrate docs from terra-scroll.
   * Migrate docs from terra-responsive-element.
 
 ## 1.12.1 - (December 14, 2021)